update java sdk test because of new pr
