+++ tests/milvus-java-test/pom.xml
@@ -118,6 +118,17 @@
             <artifactId>milvus-sdk-java</artifactId>
             <version>0.9.0-SNAPSHOT</version>
         </dependency>
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>testcontainers</artifactId>
+            <version>1.14.3</version>
+        </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-simple</artifactId>
+            <version>1.7.30</version>
+            <scope>test</scope>
+        </dependency>
 
 <!--        <dependency>-->
 <!--            <groupId>io.grpc</groupId>-->
